ABD-41: Refactor mapping methods into own class.
These methods were getting slightly too large to just live within the
model definition.

Solo: @michaelwalker
